+++ src/fios.h
@@ -100,7 +100,7 @@ void FiosFreeSavegameList();
 /* Browse to. Returns a filename w/path if we reached a file. */
 char *FiosBrowseTo(const FiosItem *item);
 /* Return path, free space and stringID */
-StringID FiosGetDescText(const char **path, uint32 *total_free);
+StringID FiosGetDescText(const char **path, uint64 *total_free);
 /* Delete a name */
 bool FiosDelete(const char *name);
 /* Make a filename from a name */
